Refactor form messages and create atoms namespace
