+++ src/components/TradeWidget/TokenRow.tsx
@@ -14,12 +14,13 @@ import Form from 'components/DepositWidget/Form'
 import { TooltipWrapper, HelpTooltipContainer, HelpTooltip } from 'components/Tooltip'
 import { InputBox } from 'components/InputBox'
 import { Input } from 'components/Input'
-import { Spinner } from 'components/Spinner'
+import { Spinner } from 'components/atoms/Spinner'
 import { WrapEtherBtn } from 'components/WrapEtherBtn'
 
 // TradeWidget: subcomponents
 import { TradeFormTokenId, TradeFormData } from 'components/TradeWidget'
-import FormMessage, { FormInputError } from 'components/TradeWidget/FormMessage'
+import { FormInputError } from 'components/atoms/FormInputError'
+import { FormMessage } from 'components/atoms/FormMessage'
 import { faPlus } from '@fortawesome/free-solid-svg-icons'
 
 // Hooks
